debug

+++ blueprints/admin_dashboard.py
@@ -92,7 +92,7 @@ def shorten():
 @app.route('/delete/<link_id>', methods=['GET'])
 def delete_link(link_id):
 
-    Link.query.filter(Link.id == link_id).delete()
+    Link.query.filter(Link.id == link_id, Link.user_id == current_user.id).delete()
     db.session.commit()
 
     return redirect(url_for('admin_dashboard.index'))
